Revert modifications from 0.5.2
The use of getResponseHeader() is causing the “Refused to get unsafe header” error in modern browsers, when the expected headers aren’t available, e.g. when there is no sso-redirect-url header.

As such I’m reverting these changes, because these errors are causing confusion.  This will break Firefox < 22, but that is a legacy browser that we will not support.